pyproject.toml
[project]
name = "alexpose"
version = "0.1.0"
description = "AI-powered gait analysis system for health condition identification"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "AlexPose Team", email = "team@alexpose.ai"}
]
license = {text = "MIT"}
keywords = ["gait-analysis", "pose-estimation", "health", "ai", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core framework dependencies
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.21",
    # AI and ML dependencies
    "openai>=1.0.0",
    "google-genai>=1.56.0",
    "mediapipe>=0.10.31",
    # Computer vision and video processing
    "opencv-python>=4.12.0.88",
    "pillow>=10.0.0",
    "yt-dlp>=2025.12.8",
    # Data processing and analysis
    "numpy>=1.24.0",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "seaborn>=0.12.0",
    # Configuration and utilities
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.1",
    "loguru>=0.7.3",
    "tenacity>=9.1.2",
    "click>=8.0.0",
    "psutil>=5.9.0",
    # Database and storage
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # Background tasks
    "celery>=5.3.0",
    "redis>=5.0.0",
    # Development and testing
    "pytest>=9.0.2",
    "hypothesis>=6.100.0",
    "ipykernel>=7.1.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "coverage>=7.13.1",
    "aiohttp>=3.13.2",
    "python-jose>=3.5.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

# Enhanced pose estimation backends
pose-enhanced = [
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

# Production deployment
production = [
    "gunicorn>=21.0.0",
    "psycopg2-binary>=2.9.0",
    "sentry-sdk[fastapi]>=1.40.0",
]

# All optional dependencies
all = [
    "alexpose[dev,pose-enhanced,production]"
]

[project.urls]
Homepage = "https://github.com/alexpose/alexpose"
Documentation = "https://docs.alexpose.ai"
Repository = "https://github.com/alexpose/alexpose"
Issues = "https://github.com/alexpose/alexpose/issues"

[project.scripts]
alexpose = "ambient.cli.main:main"
alexpose-server = "server.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ambient", "server"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["ambient", "server"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "mediapipe.*",
    "google.genai.*",
    "yt_dlp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "fast: marks tests as fast (< 1 second each)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "property: marks tests as property-based tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["ambient", "server"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
